db-management/text_detection/requirements.txt
fastapi 
uvicorn
easyocr
numpy
opencv-python
python-multipart
python-dotenv